+++ commands/help.js
@@ -13,8 +13,8 @@ module.exports.run = async (bot, message, args) => {
     .setTimestamp()
     .setFooter(message.author.tag + " | " + bot.user.username, message.author.displayAvatarURL({dynamic: true, size: 1024}))
   ;
-  if(message.channel != botCommandsChannel) {
-   message.delete()
+  if(message.channel != botCommandsChannel && message.author.id != message.guild.owner.id) {
+    message.delete()
    message.channel.send(wrongChannelEmbed).then(msg => msg.delete({timeout: 7000}));
   } else {
     let helpMembersEmbed = new Discord.MessageEmbed()
